refactor: 함수의 리턴형을 long에서 int로 수정 최대 hit은 6개이므로 int로 형변환을 해도 무방하다

+++ src/main/java/lotto/domain/Lotto.java
@@ -21,11 +21,11 @@ public class Lotto {
         return numbers.stream().sorted().toList();
     }
 
-    public long findHits(List<Integer> winningNumbers, int bonusNumber) {
+    public int findHits(List<Integer> winningNumbers, int bonusNumber) {
         long hits = numbers.stream().filter(number -> winningNumbers.stream().anyMatch(Predicate.isEqual(number)))
                 .count();
 
-        return hits;
+        return (int)hits;
     }
 
     public BonusExistence findBonus(int hits, int bonusNumber){
